Fixes #268; Solving the signedness problem with integers.

+++ xpsi/cellmesh/integrator.pyx
@@ -97,7 +97,7 @@ def integrate(size_t numThreads,
     #----------------------------------------------------------------------->>>
     cdef:
         signed int ii
-        size_t i, j, k, ks, _kdx, m, p # Array indexing
+        unsigned int i, j, k, ks, _kdx, m, p # Array indexing
         size_t T, twoT # Used globally to represent thread index
         size_t N_T = numThreads # shared
         size_t N_R = numRays # shared
